backend-render/requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS 支持
python-multipart==0.0.6

# 环境变量
python-dotenv==1.0.0

# AI 服务
openai==1.3.5  # DeepSeek API 兼容
httpx==0.25.2

# 工具库
python-dateutil==2.8.2
